Refactored into a package layout.
