--- _version.py
@@ -1,5 +1,0 @@
-# we are using the major.minor.patch notation
-# patch number increase indicates a bug fix
-# minor number increase means the addition of a feature but remaining backward compatible with earlier releases
-# major number increase indicates a significant change to the code that means it is no longer backward compatible
-__version__ = "1.01.08"
