--- db_root.py
@@ -1,1 +1,0 @@
-db_path = "%AppData%/MbClient/"
