+++ mbclient.bat
@@ -1,0 +1,2 @@
+@echo off
+python run_mbclient.py
