+++ mbclient/db_root.py
@@ -1,0 +1,1 @@
+db_path = '%AppData%/MbClient/'
