+++ mbclient/logging_config.py
@@ -1,0 +1,9 @@
+import logging
+import sys
+DEFAULT_LOG_LEVEL = logging.INFO
+
+def setup_logging(level=DEFAULT_LOG_LEVEL):
+    logging.basicConfig(level=level, format='%(asctime)s [%(levelname)s] %(name)s: %(message)s', datefmt='%Y-%m-%d %H:%M:%S', handlers=[logging.StreamHandler(sys.stdout)])
+
+def get_logger(name: str) -> logging.Logger:
+    return logging.getLogger(name)
