+++ run_mbclient.py
@@ -1,0 +1,4 @@
+from mbclient.mb_client import main
+
+if __name__ == '__main__':
+    main()
